Moving methods back to avoid circular dependency
